requirements.txt
transformers
torch
gradio
sentence-transformers
accelerate
chromadb
pypdf

# chatbot
embedchain==0.0.30
streamlit

# api
openai
langchain
fastapi
uvicorn[standard]
